Validate zTPF Memory Reserve Factor during Xmx Validation
Validate zTPF Memory Reserve Factor during Xmx Validation.
z/TPF doesn't have mmap and has a relatively small usable
memory footprint compared to linux.  When specifying an Xmx
for the z/TPF platform there is a higher chance of specifying
a higher amount then made available for the process.  This
update does a quick check during validation instead of waiting
for the runtime to detect it an run dump agents related to OOM.

Signed-off-by: James D Johnston <jjohnst@us.ibm.com>