Fix JVMTI issues related to DisposeEnvironment
The logic used to iterate JVMTI environments is incorrect in the
presence of disposed environments.  Rather than excluding disposed
environments from iteration, the old logic would terminate the iteration
entirely as soon as a disposed environment is encountered.

Add a test for SetNativeMethodPrefix (which iterates environments) in
the presence of a disposed environment.

Fixes: #2350

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>