Look for additional opportunities for stack allocation
Check for cases in which locally allocated objects are only copied
to other local variables, without values ever escaping.

These escape analysis enhancements were prototyped by Vijay
Sundarsen and Mark Stoodley.

Also made a minor improvement to getClassName to check that the node
getClassName is working with is a loadaddr for a class object;
otherwise, return it will return NULL as the class name.

Fixes:  #2072

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>