Fix recognition of hashCode function on Power
This change set fixes two issues related to recognizing the function
hashCodeImplDecompressed for String on Power.

The first issue is sometimes hashCodeImplDecompressed would get inlined.
This prevents the function from being recognized and replaced with hand
optimized assembly. This is being changed so that the function does not
get inlined if the JIT thinks it will be able to replace it with the
assembly later on.

The second issue is the hand optimized assembly for
hashCodeImplDecompressed does not work with arraylets. This was not being
checked for before and could lead to an incorrect result. This is being
changed to check if the generation of arraylets is possible. If so, the
JIT will not try to use the hand optimized assembly.

Issue: #2410
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>