Send classesThatShouldNotBe{Loaded,NewlyExtended} for CHTable commit
This was previously missed, probably because there is
also another list called classesWhichShouldNotBeNewlyExtended
which is used for a different purpose. Sending these lists
is probably required for correctness of monitor elimination
and VP but it is unknown if any current functional problems
are related to this. This change increases the number of
runtime assumptions of type 1 and 2 which are created.

Signed-off-by: Noah Weninger <noah.weninger@ibm.com>