Revert "Improve locking"
This commit reverts the first commit in
https://github.com/eclipse/openj9/pull/553.

The purpose of the original change was to reduce compare and swap
operations in the spinlock loop; thus, reducing spinning overhead. The
premise was to improve throughput by reducing spinning overhead. In the
initial benchmarking, no throughput improvements were noticed because of
this change. Recently, this change has resulted in regressing
throughput. ~30% throughput drop is observed in one of the OpenLiberty
benchmarks. Thus, it has been decided to revert the original change.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>