Add feature to ignore spin thread bound
Previously, setting lib->maxSpinThread to 0 was not allowed.

Now onwards, if lib->maxSpinThread is set to 0, then an upper bound
won't be enforced on spin threads, and unlimited spin threads will be
allowed.

Command line option, -Xthr:maxSpinThreads=<number_of_spin_threads> has
been modified to accept 0 as a valid value. Setting
-Xthr:maxSpinThreads=0 will allow unlimited spin threads.

This is an experimental feature, which is only enabled via the above
command line option. It will be used to create heuristics for locking
workloads which benefit from no bound on the number of spin threads.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>