Re-enable logCompilationSuccess call on the client
This method needs to be called on the client,
becuase it triggers hooks that need to be
triggered for jvmti agents to work correctly.

Originally, this was disabled on the client to
stop it from producing log files, but I do not see
any change in the default log files, they are still empty.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>