Skip read barriers from hardware generated guarded storage events if within a transactional region
In the event that the GSECI TX bit indicates that a hardware
generated guarded storage event occured within a transactional
execution region we should abort the read barrier and return back to
the software to abort the transaction. The reason for this is to
avoid the following scenario:

- We have a TX region in which two reference reads happen on o.f
- The first read of object o.f happens via a software read barrier
    - For example from an interpreted method
- The software read barrier will copy the object into the copy cache
- The second read of o.f.g happens via a hardware read barrier
    - This triggers a hardware transaction abort
    - All stores within the transaction are rolled back

In the above scenario before reaching the guarded storage event
handler according to Principles Of Operation S.4 the transaction
will have aborted by the time control is passed off to the event
handler. Because of this the original copy of object o.f in the copy
cache would have been rolled back to what it was prior to the GS
event handler. This location could have arbitrary data, including
a reasonable value which looks like an evacuate address which will
trip up the GC because we are effectively looking at garbage.

The solution here is to skip the read barriers from hardware
generated guarded storage event if we detect that the guarded storage
event was triggered from within a transactional execution region,
which will always result in the transaction aborting before we
reach the GS event handler.

Prior to branching back to the return address (populated by hardware
in this case to point to after the TBEGIN) we set CC=2 to mimic the
behavior of a transient transaction abort, and let the normal TABORT
software path handle the fallback.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>