Support for dynamically adjusting scratch memory space of compilations
The JIT is allowed to use up to 256 MB of 'scratch' memory for a compilation
while JSR292 compilations increase this limit to 1.5 GB. These limits create
footprint spikes during an application run which can make resource
provisioning more cumbersome.

This commit adds the following changes:
1) The JIT will use portlib APIs to compute the amount of physical memory
available (taking into consideration any container limits). This value
is cached (to minimize API overhead) and recomputed with a desired
frequency (0.5 sec by default)

2) At the beginning of a compilation the JIT checks the amount of available
physical memory and sets a limit for the scratch space accordingly. If the
amount of physical memory is so low that a relatively cheap compilation cannot
be performed, the compilation request is aborted right there.

3) When a compilation thread neeeds to allocate a new scratch memory segment
(of size 16 MB) it checks again the available physical memory. If this is low,
the compilation is aborted and will be retried at a lower optimization level.
At the same time a flag will be set which will make one compilation thread
to suspend itself (however, the JIT always makes sure there is at least one
compilation thread active).

4) If the amount of free physical memory is low, the JIT will not activate
additional compilation threads which can make the memory shortage even worse.

'low' phyical memory is determined by TR::Options::getSafeReservePhysicalMemoryValue()
which currently defaults to 50MB and can be changed with
-Xjit:safeReservePhysicalMemoryValue=<nnn>  value in KB

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>