Provide alternative API to J9::CodeCache::initialize()
Provide a version that does not take a CodeCacheHashEntrySlab parameter and
add a Doxygen comment.  This function is identical in all respects
to the version that does take a CodeCacheHashEntrySlab parameter with the
exception of that passed parameter and a rename of the codeCacheSizeAllocated
parameter to match the new OMR API.  The original version is deprecated.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>