Devirtualize `TR_PersistentCHTable::methodGotOverridden`
- `jitClassesRedefined` relies on `methodGotOverridden`
being non-virtual because it may call the method on NULL,
which doesn't result in a segfault if method is non-virtual
and doesn't use `this`.
This is a workaround, because we shouldn't be relying on undefined
behavior in the first place. I am also submitting a PR that fixes
the problem in openj9 repo.

- Add the `override` specifier to all virtual methods in
JITaaSPersistentCHTable.hpp, to ensure that overriden methods
are actually declared as virtual in the base class.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>