Clear resolveFrameFlags after each frame walk
Clear resolveFrameFlags in the interpreter stack walker after reporting
the frame. This is consistent with the JIT stack walker and avoids
leaving invalid flags set on entry to the JIT walker in very rare cases.

Also fix a small error in the resumeable stack walker, which has no
effect on the only current consumer of the feature.

Fixes: #2472

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>