Improve z/OS hardware check for Vector Extension Facility
If the `LOADxx MACHMIG VEF` load parameter is set at IPL time the user
has requested to disable the VEF support in z/OS, as such the CVTVEF
bit will not be enabled for the system.

We currently only check the LE support which says whether LE is
configured to support vector registers. In addition to this check we
also need to validate the CVTVEF bit in the CVT data structure
similarly to other features requiring OS support, such as RI, TX, or GS.

Further we take this opportunity to make improvements in the area:

- Isolate all OS checks into separate functions
  - Use them instead of directly checking CVT bits
- Add documentation on which bits we are checking
- Move checks to corresponding processor areas
- Remove stale comments

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>