JIT: Handle resolved special case invokeinterface
This commit adds jitGetImproperInterfaceMethodFromCP() so that the
compiler can query an InterfaceMethod constant pool entry to determine
whether it has been resolved to a method that does not use the usual
interface dispatch, i.e. to a private method or a method of Object. The
existing queries continue to respond in these cases as though the
constant pool entry is unresolved, so that these possibilities are only
exposed to callers prepared to deal with them.

When generating intermediate language for an invokeinterface bytecode
instruction, the compiler checks for these cases and generates a call
using the appropriate kind of dispatch. When not using interface
dispatch, it generates an explicit type check ahead of the call.

When the inliner encounters an invokeinterface bytecode instruction, it
checks for these cases in order to correctly determine which call
target(s) to consider.

Special case invokeinterface is not yet handled in cases where the
constant pool entry is unresolved at compile time. Nor is it handled in
AOT compilations, where for the time being these calls will be treated
as though their constant pool entries are unresolved.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>