JIT: Handle resolved private invokevirtual
TR_ResolvedJ9Method::getResolvedPossiblyPrivateVirtualMethod() allows
the compiler to query a VirtualMethodRef constant pool entry to
determine whether it has been resolved to a private method. The existing
getResolvedVirtualMethod() refuses to return private methods, so that
the possibility of a private method is only exposed to callers prepared
to deal with it.

When generating intermediate language for an invokevirtual bytecode
instruction, the compiler checks for a private method, and if one is
found, it generates a direct call.

When the inliner encounters an invokevirtual bytecode instruction, it
checks for a private method in order to correctly determine which call
target(s) to consider.

Private invokevirtual is not yet handled in cases where the constant
pool entry is unresolved at compile time. Nor is it handled in AOT
compilations, where for the time being these calls will be treated as
though their constant pool entries are unresolved.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>