Decode new IPIC iTable offset slot in stack walk
When encountering a JIT resolve frame for jitLookupInterfaceMethod, the
stack walk uses the itable index to determine the signature and number
of parameters of the callee. However, the IPIC slot is no longer an
itable index. It's now typically an itable offset, although it can also
be a tagged J9Method pointer or VFT offset in case of private methods or
Object methods.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>