Fix client crash when server is not running
When server is not running, `J9ClientStream` fails to establish
a connection and throws and exception, which causes a `J9Stream`
destructor to get called. Since this happens before `initStream`
was called, all of the `J9Stream` pointers are uninitialized.
Initialize all pointers to nullptr (fd to -1) in `J9Stream` constructor,
and check that they are not null in `~J9Stream`

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>