Improve interface method resolution
Remove the need for memory barriers in interface method resolution
checks by adding a special initial value for the methodIndex. The new
resolution check is that the interfaceClass is not NULL and the
methodIndex is not the special value. The special value represents a bit
pattern which can not occur in a resolved entry.

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>