Fix everything that broke after rebasing
- Add a new remote call for `getResolvedImproperInterfaceMethod` in
j9method
- In `generateInlinedCheckCastOrInstanceOfForInterface` replace the
direct call to a vm function `instanceOfOrCheckCast` with a call
to a front-end method, to correctly handle it in JITaaS.
- Remove `scanClassForReservation` from the server front-end,
and remove the remote call for this method, since it has been
deprecated in the base front end.
- Remove ClassUnload relocation, since it's not yet merged into
upstream omr, and if it's merged, it will be also merged into
upstream openj9.
- Apply a bug fix from CHTable to JITaaSCHTable (Since JITaaSCHTable
is not a child of CHTable, such chnages need to be done manually).
- Send rtResolve runtime flag from client to server.
This bug was revealed by new code, sending just one flag is a
temporary solution, until we come up with a better way.
- Fix a bug in `JITaaSIProfiler::searchForMethodSample`, string read
from the stream needs to be copied, not referenced.
- Other minor fixes of things that broke after rebasing

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>