Prevent class initialization check bypass
If the JIT encounters an unresolved static field at compile-time, the
generated resolve code calls jitResolveClassFromStaticField to resolve
the declaring class of the field. This code was previously filling in
the RAM constant pool entry for the static field, allowing other threads
to bypass the class initialization check, resulting in access to the
static field before the class is fully initialized.

The fix is to resolve the static field into local data instead of the
constant pool. This call will be immediately followed up by the
jitResolveStaticField or jitResolveStaticFieldSetter, which will fully
resolve the constant pool entry in the expected case.

Refactor field resolve checks into helpers which can be used everywhere.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>