Add Heuristic Region Infrastructure
The Symbol Validation Manager is used to remember all the different ways
the compiler acquired a symbol. However, there are regions where this
information isn't necessary for the load run. For example, when asking
for something from the Persistent CH Table, all that matters is the
input to the query and the output; however, the query involves going
through multiple classes before finally returning the answer. If the
compiler were to store valiations for everything that occurs during this
query, it would require consistency between the compile and load run for
more than what is only required for the compilation.

Therefore, this commit adds the means to denote regions that are not
strictly necessary to validate for the load run.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>