Enable support for metronome gc policy
When client uses metronome gc policy, need to generate spine checks
for discontiguous arrays. All information that server needs to generate
correct code is captured in 3 variables of J9ObjectModel.

To store these variables, and all other variables that are set once
per client
and never change, create a new struct in `ClientSessionData` - `VMInfo`.
Whenever server needs to get one of the values in that struct, it calls
`getOrCacheVMInfo`, which makes a remote call and fetches all the values
the first time its called, and retrieves the requested variable from
cache in all subsequent calls.

If more persistent values per client are found, they should be added to
the `VMInfo` struct.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>