Add TR_J2IVirtualThunkPointer relocation kind
This relocates a J2I (JIT to interpreter) virtual thunk pointer along
with a nearby constant pool address and index. The relocation is
necessary because with nestmates, the data consumed by PicBuilder now
contains a J2I thunk pointer, which is used for direct dispatch in case
the callee has not yet been compiled.

No such relocations are created yet. That will be done in each code
generator's call snippet code.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>