Fix linker (ld) cannot find omrsig on ARM
On ARM, j9ddrgen and constgen are build using the -static compiler flag.
-static prevents shared libraries to be linked.

When J9VM_PORT_OMRSIG_SUPPORT is enabled, j9ddrgen and constgen need to
link to omrsig, which is a shared library.

Now onwards, -static compiler flag will not be used for j9ddrgen and
constgen on ARM if J9VM_PORT_OMRSIG_SUPPORT is enabled. This will allow
j9ddrgen and constgen to link to omrsig.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>