Use boolean expressions in if statements
While compiling OpenJ9 with Clang 9.1.0 on OSX, the following compile
error is encountered:

CompositeCacheTest.cpp:251:13: error: using the result of an assignment
as a condition without parentheses [-Werror,-Wparentheses]
        if (result |= localTestStats(
            ~~~~~~~^~~~~~~~~~~~~~~~~~~
CompositeCacheTest.cpp:251:13: note: place parentheses around the
assignment to silence this warning
        if (result |= localTestStats(
                   ^
CompositeCacheTest.cpp:251:13: note: use '!=' to turn this compound
assignment into an inequality comparison
        if (result |= localTestStats(
                   ^~
                   !=

The above error happens because the output of an assignment is used in
the if statements instead of a boolean expression. Boolean expressions
are used within the if statements to fix the above error.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>