Set PPCVirtualUnresolvedSnippet as GCSafePoint
The recently added changes to PPCVirtualUnresolvedSnippet to support
private virtual calls now require the snippet to be considered a GC Safe
Point. This is necessary to generate GC map information.

The PPCVirtualUnresolvedSnippet constructor was changed so that it now
passed true for the value of isGCSafePoint to the parent constructor.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>