Fix value of j9binBuffer/jrebinBuffer [OSX]
On OSX, j9binBuffer and jrebinBuffer don't point to the compressedrefs
directory. This prevents jvm.c::preloadLibrary to derive correct paths
to libraries such as omrsig, j9vm* etc.

j9binBuffer and jrebinBuffer are derived using jvmDLLNameBuffer.
jvmDLLNameBuffer relies on the path to libj9vm*.so/dll/dylib. On OSX,
incorrect suffix is used for libj9vm*.<SUFFIX>. This causes j9binBuffer
and jrebinBuffer to be incorrectly set and jvm.c::preloadLibrary to
malfunction.

This fix will correctly set the library suffix for libj9vm* on OSX.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>