Handle 8-byte call in VPIC snippet for nestmates
When x86 PicBuilder is carrying out a direct call to a private method of
a nestmate, it sets the return address so that the call will return into
the code in the VPIC snippet, which in the case of a virtual dispatch
would be something like the following:

    call qword ptr [$vftReg+$vftOffset]
    jmp mainline

In particular, the callee should return to the jmp instruction, whose
address has been calculated as 7 + the address of the call instruction.
But the call is sometimes one byte longer. The variability is now dealt
with as follows:

- When overwriting the call, PicBuilder now writes a value that works
  for both 7- and 8-byte offsets to the jmp instruction.

- When doing direct dispatch for nestmates, PicBuilder now detects the
  offset to jmp, in order to compute the correct return address.

Fixes #2955
Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>