Use new OMR variables for installation locations
OMR is removing CMake install logic that is very specific to the OpenJ9
build system, in favor of a customizable configuration.  To accomplish
this, OMR is introducing new variables to customize the installation
directories of OMR artifacts.

OpenJ9 uses the install target in OMR to stage certain artifacts, such
as omrsig and tracegen, into the build tree. The end result is
functionally the same from the OpenJ9 perspective.

Signed-off-by: Andrew Young <youngar17@gmail.com>