Create remote calls to get constant strings from resolved method
Make remote calls for `isUnresolvedString` and `constantString`.
This is needed to correctly determine whether a String sym reference
is resolved. Previously, `isUnresolvedString` was always returning true,
which caused all corresponding sym references to be marked as
unresolved. This caused some blocks to be incorrectly marked as cold,
which prevented inlinining and caused a throughput regression.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>