Introduce CALL_C_ADDR_WITH_VMTHREAD for address calls
When the macro CALL_C_WITH_VMTHREAD is used to make a call to an
address, it results in an underscore being prepended to qword
on OSX, forming an invalid token. This new macro acts as an extra
layer to fix this issue, where it gets calls the address correctly
on OSX, and gets defined as CALL_C_WITH_VMTHREAD on Windows and
Linux.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>