Eliminate remote calls for isWarmCallGraphTooBig
Previously, calls to `isWarmCallGraphTooBig`
always made remote calls to the client, which used IProfiler to
determine
whether the callgraph is too big to be inlined from the bytecode entry.
However, since we are already caching IProfiler bytecode entries on the
server, we can just access the cached result.
We ensure that `isWarmCallGraphTooBig` always returns the correct value
by making `setWarmCallGraphTooBig` update the value in cached
entries, as well as on the client. This way, even if a current server
crashes, the new server will still receive correct values.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>