Fix Class.toGenericString array output on jdk9+
Java 9 and up have different output for toGenericString for array
classes in order to be more human readable instead of using encodings
like '[[Ljava.lang.string'. This commit changes the implementation in
order to match the new standard for Java 9 and up.

Also change tests for Java 9 and up to use the new output format.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>