Add AOT Validations
This commit adds uses of the new Symbol Validation Manager infrastructure
to do AOT validations (under an option). It also introduces the
following two changes:

1. Use resolved interpreter dispatch for AOT

Previously, because almost everything was "unresolved" under AOT, all
calls would use the unresolved dispatch. However, with the new
validations, very few unresolved pointers are only unresolved because of
AOT. Therefore, generating an unresolved dispatch is incorrect.

The solution to this problem is to assume, during an AOT compile, that
all methods are interpreted. Thus, if a method is resolved, the compiler
will generate a resolved interpreter dispatch snippet, which can patch
the caller if/when the method is compiled.

2. Fix ObjectAllocationInlining under AOT

Previously, Object Allocation Inlining under AOT maintained certain
assumtpions about the trees that was only true if the class was
unresolved. This commit fixes those assumptions.

Signed-off-by: Irwin D'Souza