Fix failures in Condy Test
When GC kicks in when constant dynamic object is resolved, VM helper
code expects all the registers are stored on Java Stack and while
resolving frames it takes this adjustment into account while building
frames. Condy Test was failing as it was not storing registers on Java
Stack producing corrupted stack frame.

Fixes: #3082

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>