Change heuristics regarding AOT compilations
With the current heuristics only methods that have been downgraded from warm
to cold optimization level are eligible for AOT compilations. Moreover,
methods that are deemed "important for start-up" (string/zip/hash) are not
downgraded and therefore not AOTed.

This commit changes the heuristics such that the methods "important for
start-up" become subject to the downgrading policy (like any other method)
and therefore can be AOTed. However, when generating an AOT body for such
an "important method for start-up" the optimization level is raised back
to warm.

Performance experiments with Liberty Applicatioon Server show a 5-6% start-up
time improvement from this change.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>